refactor(profiles): split ProfileManager's live half into an excluded companion
The AppState-bound orchestration — performSetup, applyProfile and its
snapshot push, profile hotkey registration, display auto-switch, and the
Focus Filter handlers — moves verbatim to ProfileManager+Live.swift. None
of it can run in a unit test: it needs a running AppState, WindowServer
display identity via Bridging, Carbon hotkey registration, and Focus
Filter intents.

Capture now takes the three stores it actually reads (settings,
appearance manager, item manager) instead of the full AppState — the same
seam updateProfileLayout(id:itemManager:) already used — so the
save/update paths stay measured and testable. The AppState overloads
forward from the companion with unchanged signatures; no call sites move.

A handful of members widen from private to internal because an extension
in another file cannot reach private members; the declaration site notes
why.
